fix(tokenizer): SHA-1 integrity gate + split CI so pure-Go loader tests run
- bpe_loader.go: verify the on-disk cl100k table against the SHA-1 of its
  *contents* and refuse to load a corrupt/tampered file (mismatch returns an
  error, consistent with the malformed-table path). The pinned digest is the
  file-content SHA-1 (6494e42d...), not the tiktoken cache filename
  (223921b76... = sha1(bpeURL)), which previously would have rejected the
  genuine table.
- bpe_loader_test.go: writeBpeTable registers the synthetic table's SHA-1 so
  the gate accepts it while still exercising the verification path.
- Split CI tiers: tag the C++-binding / dict-dependent tests
  (tokenizer_test.go, tokenizer_concurrent_test.go) and the on-disk anchor
  test (bpe_loader_anchors_test.go) with `//go:build manual` so they only run
  under the docker builder's /usr/share/infinity/resource. Remove the
  `internal/tokenizer` exclusion from tests.yml and sep-tests.yml so the
  pure-Go BPE loader tests (bpe_loader_test.go) run in the default tier.
