+++ internal/tokenizer/tokenizer_concurrent_test.go
@@ -12,7 +12,7 @@
 //  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 //  See the License for the specific language governing permissions and
 //  limitations under the License.
-//
+//go:build manual
 
 package tokenizer
